+++ src/main/java/com/threejavers/drinkapp/dto/CategoryDTO.java
@@ -1,4 +1,4 @@
-package com.threejavers.drinkapp.DTO;
+package com.threejavers.drinkapp.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -9,9 +9,8 @@ import java.util.UUID;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class IngredientDTO {
+public class CategoryDTO {
 
     private UUID id;
     private String name;
-    private String measure;
 }
